move ConanApp to a different module
